Rename particle system files
