+++ include/overlay011/ov11_0221F840.h
@@ -1,7 +1,7 @@
 #ifndef POKEPLATINUM_OV11_0221F840_H
 #define POKEPLATINUM_OV11_0221F840_H
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 ParticleSystem *ov11_0221F840(int heapID);
 ParticleSystem *ov11_0221F888(int heapID);
