+++ include/overlay104/ov104_0223D5D0.h
@@ -1,7 +1,7 @@
 #ifndef POKEPLATINUM_OV104_0223D5D0_H
 #define POKEPLATINUM_OV104_0223D5D0_H
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 #include "overlay104/struct_ov104_0223D5F0_decl.h"
