+++ include/struct_defs/struct_0207C8C4.h
@@ -1,7 +1,7 @@
 #ifndef POKEPLATINUM_STRUCT_0207C8C4_H
 #define POKEPLATINUM_STRUCT_0207C8C4_H
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "struct_defs/struct_0207C894.h"
 
 #include "spl.h"
