+++ src/battle/ov16_0223B140.c
@@ -73,7 +73,7 @@
 #include "trainer_info.h"
 #include "unk_0200C440.h"
 #include "unk_0200F174.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0201567C.h"
 #include "unk_02015F84.h"
 #include "unk_0202419C.h"
