+++ src/overlay010/ov10_0221F800.c
@@ -6,7 +6,7 @@
 #include "struct_decls/font_oam.h"
 #include "struct_decls/struct_0200C440_decl.h"
 #include "struct_decls/struct_02012744_decl.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "struct_defs/struct_020127E8.h"
 #include "struct_defs/struct_0207C690.h"
 #include "struct_defs/struct_02099F80.h"
@@ -50,7 +50,7 @@
 #include "unk_0200C440.h"
 #include "unk_0200F174.h"
 #include "unk_02012744.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202419C.h"
 #include "unk_02024220.h"
 #include "unk_0202F1D4.h"
