+++ src/overlay012/ov12_022237EC.c
@@ -3,7 +3,7 @@
 #include <nitro.h>
 #include <string.h>
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 #include "overlay011/ov11_0221F840.h"
 #include "overlay012/ov12_0222389C.h"
@@ -11,7 +11,7 @@
 
 #include "narc.h"
 #include "spl.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 static void ov12_022237EC(SPLEmitter *param0);
