+++ src/overlay012/ov12_02225640.c
@@ -10,7 +10,7 @@
 
 #include "camera.h"
 #include "spl.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 static void ov12_02225640(UnkStruct_ov12_02225640 *param0, int param1[]);
 static BOOL ov12_0222564C(SPLEmitter *param0, UnkStruct_ov12_02225640 *param1);
