+++ src/overlay012/ov12_02235254.c
@@ -3,7 +3,7 @@
 #include <nitro.h>
 #include <string.h>
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 #include "overlay012/ov12_0221FC20.h"
 #include "overlay012/struct_ov12_0221FCDC_decl.h"
@@ -15,7 +15,7 @@
 #include "inlines.h"
 #include "pokemon_sprite.h"
 #include "sprite_system.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 void ov17_022413D8(void);
 void include_unk_ov12_0223A218(VecFx32 *dummy);
