+++ src/overlay012/ov12_02235E94.c
@@ -3,7 +3,7 @@
 #include <nitro.h>
 #include <string.h>
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "struct_defs/struct_0202CA28.h"
 #include "struct_defs/struct_0202CA64.h"
 
@@ -32,7 +32,7 @@
 #include "sprite_util.h"
 #include "sys_task.h"
 #include "sys_task_manager.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202C9F4.h"
 #include "unk_02097B18.h"
