+++ src/overlay017/ov17_0223CB1C.c
@@ -38,7 +38,7 @@
 #include "system.h"
 #include "touch_pad.h"
 #include "unk_0200F174.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_020366A0.h"
 #include "unk_020393C8.h"
 #include "unk_020933F8.h"
