+++ src/overlay017/ov17_0223DAD0.c
@@ -47,7 +47,7 @@
 #include "touch_pad.h"
 #include "unk_0200F174.h"
 #include "unk_02012744.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_020366A0.h"
 #include "unk_020393C8.h"
 #include "unk_020933F8.h"
