+++ src/overlay077/ov77_021D25B0.c
@@ -33,7 +33,7 @@
 #include "sys_task_manager.h"
 #include "system.h"
 #include "unk_0200F174.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202419C.h"
 #include "unk_02024220.h"
 #include "vram_transfer.h"
