+++ src/overlay077/ov77_021D6C70.c
@@ -3,7 +3,7 @@
 #include <nitro.h>
 #include <string.h>
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 #include "camera.h"
 #include "heap.h"
@@ -11,7 +11,7 @@
 #include "pokemon_sprite.h"
 #include "rtc.h"
 #include "spl.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202419C.h"
 
 typedef struct UnkStruct_ov77_021D6CFC_t {
