+++ src/overlay117/ov117_02260668.c
@@ -50,7 +50,7 @@
 #include "trainer_info.h"
 #include "unk_0200F174.h"
 #include "unk_02012744.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202419C.h"
 #include "unk_02024220.h"
 #include "unk_020363E8.h"
