+++ src/particle_system.c
@@ -1,4 +1,4 @@
-#include "unk_02014000.h"
+#include "particle_system.h"
 
 #include <nitro.h>
 #include <string.h>
