+++ src/unk_0207C63C.c
@@ -2,7 +2,7 @@
 #include <nnsys.h>
 #include <string.h>
 
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "struct_defs/struct_0207C690.h"
 #include "struct_defs/struct_0207C894.h"
 #include "struct_defs/struct_0207C8C4.h"
@@ -11,7 +11,7 @@
 #include "gx_layers.h"
 #include "heap.h"
 #include "spl.h"
-#include "unk_02014000.h"
+#include "particle_system.h"
 #include "unk_0202419C.h"
 #include "unk_02024220.h"
